OpenCorePkg/Library/OcCompressionLib: fix signed/unsigned compare

+++ Library/OcCompressionLib/OcCompressionLib.c
@@ -93,7 +93,7 @@ DecompressMaskedRLE24 (
         //
         // Early exit on not enough bytes to read or fill.
         //
-        if ((UINT32) (SrcEnd - Src) < ControlValue || DstEnd - DstCur < ControlValue * sizeof (UINT32)) {
+        if ((UINT32) (SrcEnd - Src) < ControlValue || (UINT32) (DstEnd - DstCur) < ControlValue * sizeof (UINT32)) {
           return 0;
         }
